requirements.txt
aiogram>=3.0.0
openai>=1.3.5
sentence-transformers>=2.2.2
sqlalchemy>=2.0.0
gspread-asyncio>=1.2.0
oauth2client>=4.1.3
pydantic>=1.10.0
pydantic-settings>=2.10.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
numpy>=1.26.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
chromadb>=0.4.15
onnxruntime>=1.20.0